% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicDataSet.R
\name{create_multiOmicDataSet_from_dataframes}
\alias{create_multiOmicDataSet_from_dataframes}
\title{Construct a multiOmicDataSet object from data frames}
\usage{
create_multiOmicDataSet_from_dataframes(
  sample_metadata,
  counts_dat,
  sample_id_colname = NULL,
  feature_id_colname = NULL,
  count_type = "raw"
)
}
\arguments{
\item{sample_metadata}{sample metadata as a data frame or tibble. The first column is assumed to contain the sample
IDs which must correspond to column names in the raw counts.}

\item{counts_dat}{data frame of feature counts (e.g. expected feature counts from RSEM).}

\item{sample_id_colname}{name of the column in \code{sample_metadata} that contains the sample IDs. (Default: \code{NULL} -
first column in the sample metadata will be used.)}

\item{feature_id_colname}{name of the column in \code{counts_dat} that contains feature/gene IDs. (Default: \code{NULL} - first
column in the count data will be used.)}

\item{count_type}{type to assign the values of \code{counts_dat} to in the \code{counts} slot}
}
\value{
A \code{multiOmicDataSet} object.
}
\description{
Construct a multiOmicDataSet object from data frames
}
\examples{

base_url <- paste0(
  "https://raw.githubusercontent.com/CCBR/MOSuite/",
  "refs/tags/v0.4.2/inst/extdata/nidap/"
)
sample_meta_dat <- readr::read_csv(paste0(
  base_url, "Sample_Metadata_Bulk_RNA-seq_Training_Dataset_CCBR.csv.gz"
))
feature_counts_dat <- readr::read_csv(paste0(base_url, "Raw_Counts.csv.gz"))
moo <- create_multiOmicDataSet_from_dataframes(
  sample_metadata = sample_meta_dat,
  counts_dat = feature_counts_dat
)
head(moo@sample_meta)
head(moo@counts[['raw']])

}
