% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicDataSet.R
\name{create_multiOmicDataSet_from_files}
\alias{create_multiOmicDataSet_from_files}
\title{Construct a multiOmicDataSet object from text files (e.g. TSV, CSV).}
\usage{
create_multiOmicDataSet_from_files(
  sample_meta_filepath,
  feature_counts_filepath,
  count_type = "raw",
  sample_id_colname = NULL,
  feature_id_colname = NULL,
  delim = NULL,
  ...
)
}
\arguments{
\item{sample_meta_filepath}{path to text file with sample IDs and metadata for differential analysis.}

\item{feature_counts_filepath}{path to text file of expected feature counts (e.g. gene counts from RSEM).}

\item{count_type}{type to assign the values of \code{counts_dat} to in the \code{counts} slot}

\item{sample_id_colname}{name of the column in \code{sample_metadata} that contains the sample IDs. (Default: \code{NULL} -
first column in the sample metadata will be used.)}

\item{feature_id_colname}{name of the column in \code{counts_dat} that contains feature/gene IDs. (Default: \code{NULL} - first
column in the count data will be used.)}

\item{delim}{Delimiter used in the input files. Any delimiter accepted by \code{readr::read_delim()} can be used.
If the files are in CSV format, set \code{delim = ','}; for TSV format, set \code{delim = '\\t'}.}

\item{...}{additional arguments forwarded to \code{readr::read_delim()}.}
}
\value{
A \code{multiOmicDataSet} object.
}
\description{
Construct a multiOmicDataSet object from text files (e.g. TSV, CSV).
}
\examples{
base_url <- paste0(
  "https://raw.githubusercontent.com/CCBR/MOSuite/",
  "refs/tags/v0.4.2/inst/extdata/nidap/"
)
moo <- create_multiOmicDataSet_from_files(
  sample_meta_filepath = paste0(
    base_url, "Sample_Metadata_Bulk_RNA-seq_Training_Dataset_CCBR.csv.gz"
  ),
  feature_counts_filepath = paste0(base_url, "Raw_Counts.csv.gz")
)
head(moo@sample_meta)
head(moo@counts[['raw']])

}
