% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicDataSet.R
\name{extract_counts}
\alias{extract_counts}
\alias{extract_counts,MOObject::multiOmicDataSet-method}
\title{Extract count data}
\usage{
extract_counts(moo, count_type, sub_count_type = NULL)

## S7 method for class <MOObject::multiOmicDataSet>
extract_counts(moo, count_type, sub_count_type = NULL)
}
\arguments{
\item{moo}{multiOmicDataSet containing \code{count_type} & \code{sub_count_type} in the counts slot}

\item{count_type}{the type of counts to use -- must be a name in the counts slot (\code{moo@counts[[count_type]]})}

\item{sub_count_type}{if \code{count_type} is a list, specify the sub count type within the list
(\code{moo@counts[[count_type]][[sub_count_type]]}). (Default: \code{NULL})}
}
\value{
A data frame of counts.
}
\description{
Extract count data
}
\examples{
sample_metadata <- data.frame(sample_id = c("s1", "s2"), group = c("A", "B"))
counts_dat <- data.frame(
  feature_id = c("gene1", "gene2"),
  s1 = c(10, 20),
  s2 = c(15, 25)
)
moo <- create_multiOmicDataSet_from_dataframes(sample_metadata, counts_dat)
extract_counts(moo, "raw")
}
\seealso{
Other moo:
\code{\link[=create_multiOmicDataSet_from_dataframes]{create_multiOmicDataSet_from_dataframes()}},
\code{\link[=create_multiOmicDataSet_from_files]{create_multiOmicDataSet_from_files()}},
\code{\link[=multiOmicDataSet]{multiOmicDataSet()}},
\code{\link[=read_multiOmicDataSet]{read_multiOmicDataSet()}},
\code{\link[=read_multiOmicDataSet_properties]{read_multiOmicDataSet_properties()}},
\code{\link[=write_multiOmicDataSet]{write_multiOmicDataSet()}},
\code{\link[=write_multiOmicDataSet_properties]{write_multiOmicDataSet_properties()}}
}
\concept{moo}
