% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOObject-package.R
\docType{package}
\name{MOObject-package}
\alias{MOObject}
\alias{MOObject-package}
\title{MOObject: multiOmicDataSet object class for MOSuite}
\description{
MOObject defines the multiOmicDataSet object class which is used in MOSuite for differential RNA-seq and multi-omics analyses. Separating the MOObject class definition from MOSuite allows for lightweight installations in downstream workflows. See the website for more information: \url{https://ccbr.github.io/MOObject/}. See the MOSuite website for the most comprehensive documentation: \url{https://ccbr.github.io/MOSuite/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CCBR/MOObject}
  \item \url{https://ccbr.github.io/MOObject/}
  \item Report bugs at \url{https://github.com/CCBR/MOObject/issues}
}

}
\author{
\strong{Maintainer}: Kelly Sovacool \email{kelly.sovacool@nih.gov} (\href{https://orcid.org/0000-0003-3283-829X}{ORCID})

Authors:
\itemize{
  \item Kelly Sovacool \email{kelly.sovacool@nih.gov} (\href{https://orcid.org/0000-0003-3283-829X}{ORCID})
  \item Philip Homan \email{philip.homan@nih.gov}
  \item Vishal Koparde \email{vishal.koparde@nih.gov} (\href{https://orcid.org/0000-0001-8978-8495}{ORCID})
}

Other contributors:
\itemize{
  \item CCR Collaborative Bioinformatics Resource [copyright holder]
}

}
\keyword{internal}
