% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicDataSet.R
\name{read_analyses_dir_}
\alias{read_analyses_dir_}
\title{Read analyses from a directory written by write_multiOmicDataSet_properties}
\usage{
read_analyses_dir_(analyses_dir)
}
\arguments{
\item{analyses_dir}{Path to the \verb{analyses/} subdirectory.}
}
\value{
A named list of data frames, lists, or arbitrary R objects.
}
\description{
Read analyses from a directory written by write_multiOmicDataSet_properties
}
\keyword{internal}
