% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicDataSet.R
\name{read_counts_dir_}
\alias{read_counts_dir_}
\title{Read counts from a directory written by write_multiOmicDataSet_properties}
\usage{
read_counts_dir_(counts_dir)
}
\arguments{
\item{counts_dir}{Path to the \verb{counts/} subdirectory.}
}
\value{
A named list of data frames (and nested lists for sub-count types).
}
\description{
Read counts from a directory written by write_multiOmicDataSet_properties
}
\keyword{internal}
