% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicDataSet.R
\name{write_multiOmicDataSet}
\alias{write_multiOmicDataSet}
\title{Write a multiOmicDataSet to RDS}
\usage{
write_multiOmicDataSet(moo, filepath = "moo.rds")
}
\arguments{
\item{moo}{\link{multiOmicDataSet} object to serialize}

\item{filepath}{Path to the RDS file to write (default: "moo.rds")}
}
\value{
Invisibly returns \code{filepath}.
}
\description{
Write a multiOmicDataSet to RDS
}
