% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicDataSet.R
\name{write_multiOmicDataSet_properties}
\alias{write_multiOmicDataSet_properties}
\alias{write_multiOmicDataSet_properties,MOObject::multiOmicDataSet-method}
\title{Write multiOmicDataSet properties to disk}
\usage{
write_multiOmicDataSet_properties(moo, output_dir = "moo")

## S7 method for class <MOObject::multiOmicDataSet>
write_multiOmicDataSet_properties(moo, output_dir = "moo")
}
\arguments{
\item{moo}{multiOmicDataSet object to write.}

\item{output_dir}{output directory.}
}
\value{
Invisibly returns \code{output_dir}.
}
\description{
Write multiOmicDataSet properties to disk
}
