% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicDataSet.R
\name{create_multiOmicDataSet_from_dataframes}
\alias{create_multiOmicDataSet_from_dataframes}
\title{Construct a multiOmicDataSet object from data frames}
\usage{
create_multiOmicDataSet_from_dataframes(
  sample_metadata,
  counts_dat,
  sample_id_colname = NULL,
  feature_id_colname = NULL,
  count_type = "raw"
)
}
\arguments{
\item{sample_metadata}{sample metadata as a data frame or tibble. The first column is assumed to contain the sample
IDs which must correspond to column names in the raw counts.}

\item{counts_dat}{data frame of feature counts (e.g. expected feature counts from RSEM).}

\item{sample_id_colname}{name of the column in \code{sample_metadata} that contains the sample IDs. (Default: \code{NULL} -
first column in the sample metadata will be used.)}

\item{feature_id_colname}{name of the column in \code{counts_dat} that contains feature/gene IDs. (Default: \code{NULL} - first
column in the count data will be used.)}

\item{count_type}{type to assign the values of \code{counts_dat} to in the \code{counts} slot}
}
\value{
A \link{multiOmicDataSet} object.
}
\description{
Construct a multiOmicDataSet object from data frames
}
\examples{

sample_meta_dat <- data.frame(sample_id = c("s1", "s2"), group = c("A", "B"))
feature_counts_dat <- data.frame(
  feature_id = c("gene1", "gene2"),
  s1 = c(10, 20),
  s2 = c(15, 25)
)
moo <- create_multiOmicDataSet_from_dataframes(
  sample_metadata = sample_meta_dat,
  counts_dat = feature_counts_dat
)
head(moo@sample_meta)
head(moo@counts[['raw']])

}
\seealso{
Other moo:
\code{\link[=create_multiOmicDataSet_from_files]{create_multiOmicDataSet_from_files()}},
\code{\link[=extract_counts]{extract_counts()}},
\code{\link[=multiOmicDataSet]{multiOmicDataSet()}},
\code{\link[=read_multiOmicDataSet]{read_multiOmicDataSet()}},
\code{\link[=read_multiOmicDataSet_properties]{read_multiOmicDataSet_properties()}},
\code{\link[=write_multiOmicDataSet]{write_multiOmicDataSet()}},
\code{\link[=write_multiOmicDataSet_properties]{write_multiOmicDataSet_properties()}}
}
\concept{moo}
