% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicDataSet.R
\name{create_multiOmicDataSet_from_files}
\alias{create_multiOmicDataSet_from_files}
\title{Construct a multiOmicDataSet object from delimited files}
\usage{
create_multiOmicDataSet_from_files(
  sample_meta_filepath,
  feature_counts_filepath,
  count_type = "raw",
  sample_id_colname = NULL,
  feature_id_colname = NULL,
  delim = NULL,
  ...
)
}
\arguments{
\item{sample_meta_filepath}{path to sample metadata file.}

\item{feature_counts_filepath}{path to feature counts file.}

\item{count_type}{type to assign the values of \code{counts_dat} to in \code{counts}.}

\item{sample_id_colname}{column in \code{sample_metadata} that contains sample IDs.
If \code{NULL}, use the first column.}

\item{feature_id_colname}{column in \code{counts_dat} that contains feature IDs.
If \code{NULL}, use the first column.}

\item{delim}{delimiter for \code{readr::read_delim()}.}

\item{...}{additional arguments passed to \code{readr::read_delim()}.}
}
\value{
A \code{multiOmicDataSet} object.
}
\description{
Construct a multiOmicDataSet object from delimited files
}
