% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicDataSet.R
\name{extract_counts}
\alias{extract_counts}
\alias{extract_counts,MOObject::multiOmicDataSet-method}
\title{Extract count data from a multiOmicDataSet}
\usage{
extract_counts(moo, count_type, sub_count_type = NULL)

## S7 method for class <MOObject::multiOmicDataSet>
extract_counts(moo, count_type, sub_count_type = NULL)
}
\arguments{
\item{moo}{multiOmicDataSet containing counts.}

\item{count_type}{count type in \code{moo@counts}.}

\item{sub_count_type}{subtype in \code{moo@counts[[count_type]]} when that entry is a list.}
}
\value{
A data frame of counts.
}
\description{
Extract count data from a multiOmicDataSet
}
