% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicDataSet.R
\name{extract_counts}
\alias{extract_counts}
\alias{extract_counts,MOObject::multiOmicDataSet-method}
\title{Extract count data}
\usage{
extract_counts(moo, count_type, sub_count_type = NULL)

## S7 method for class <MOObject::multiOmicDataSet>
extract_counts(moo, count_type, sub_count_type = NULL)
}
\arguments{
\item{moo}{multiOmicDataSet containing \code{count_type} & \code{sub_count_type} in the counts slot}

\item{count_type}{the type of counts to use -- must be a name in the counts slot (\code{moo@counts[[count_type]]})}

\item{sub_count_type}{if \code{count_type} is a list, specify the sub count type within the list
(\code{moo@counts[[count_type]][[sub_count_type]]}). (Default: \code{NULL})}
}
\value{
A data frame of counts.
}
\description{
Extract count data
}
