% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicDataSet.R
\name{multiOmicDataSet}
\alias{multiOmicDataSet}
\title{multiOmicDataSet class}
\usage{
multiOmicDataSet(sample_metadata, anno_dat, counts_lst, analyses_lst = list())
}
\arguments{
\item{sample_metadata}{sample metadata as a data frame. The first column is
assumed to contain sample IDs that correspond to columns in raw counts.}

\item{anno_dat}{data frame of feature annotations.}

\item{counts_lst}{named list of count data frames. Each count data frame must
have a feature ID column first and sample columns after.}

\item{analyses_lst}{named list of analysis results.}
}
\value{
A \code{multiOmicDataSet} S7 object.
}
\description{
multiOmicDataSet class
}
\section{Additional properties}{

\describe{
\item{\code{@sample_meta}}{sample metadata as a data frame.}

\item{\code{@annotation}}{feature annotation data frame.}

\item{\code{@counts}}{named list of counts data frames.}

\item{\code{@analyses}}{named list of analysis results.}
}}

