% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicDataSet.R
\name{multiOmicDataSet}
\alias{multiOmicDataSet}
\title{multiOmicDataSet class}
\usage{
multiOmicDataSet(sample_metadata, anno_dat, counts_lst, analyses_lst = list())
}
\arguments{
\item{sample_metadata}{sample metadata as a data frame or tibble. The first column is assumed to contain the sample
IDs which must correspond to column names in the raw counts.}

\item{anno_dat}{data frame of feature annotations, such as gene symbols or any other information about the features
in \code{counts_lst}.}

\item{counts_lst}{named list of data frames containing counts, e.g. expected feature counts from RSEM. Each data
frame is expected to contain a \code{feature_id} column as the first column, and all remaining columns are sample IDs in
the \code{sample_meta}.}

\item{analyses_lst}{named list of analysis results, e.g. DESeq results object}
}
\value{
A \link{multiOmicDataSet} S7 object.
}
\description{
multiOmicDataSet class
}
\examples{
sample_metadata <- data.frame(sample_id = c("s1", "s2"), group = c("A", "B"))
anno_dat <- data.frame(feature_id = c("gene1", "gene2"))
counts_dat <- data.frame(
  feature_id = c("gene1", "gene2"),
  s1 = c(10, 20),
  s2 = c(15, 25)
)
moo <- multiOmicDataSet(
  sample_metadata = sample_metadata,
  anno_dat = anno_dat,
  counts_lst = list(raw = counts_dat)
)
moo
}
\seealso{
Other moo:
\code{\link[=create_multiOmicDataSet_from_dataframes]{create_multiOmicDataSet_from_dataframes()}},
\code{\link[=create_multiOmicDataSet_from_files]{create_multiOmicDataSet_from_files()}},
\code{\link[=extract_counts]{extract_counts()}},
\code{\link[=read_multiOmicDataSet]{read_multiOmicDataSet()}},
\code{\link[=read_multiOmicDataSet_properties]{read_multiOmicDataSet_properties()}},
\code{\link[=write_multiOmicDataSet]{write_multiOmicDataSet()}},
\code{\link[=write_multiOmicDataSet_properties]{write_multiOmicDataSet_properties()}}
}
\concept{moo}
\section{Additional properties}{

\describe{
\item{\code{@sample_meta}}{sample metadata as a data frame or tibble. The first column is assumed to contain the sample
IDs which must correspond to column names in the raw counts.}

\item{\code{@annotation}}{data frame of feature annotations, such as gene symbols or any other information about the
features in the counts list.}

\item{\code{@counts}}{named list of counts data frames (e.g. \code{raw}, \code{clean}, \code{cpm}, \code{filt}, \code{norm}, \code{batch}). Each data
frame is expected to contain a feature ID column as the first column, and all remaining columns are sample IDs.}

\item{\code{@analyses}}{named list of analysis results (e.g. DESeq2 results, colors).}
}}

