% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicDataSet.R
\name{read_multiOmicDataSet}
\alias{read_multiOmicDataSet}
\title{Read a multiOmicDataSet from disk}
\usage{
read_multiOmicDataSet(filepath)
}
\arguments{
\item{filepath}{path to an RDS file produced by \code{\link[=write_multiOmicDataSet]{write_multiOmicDataSet()}}.}
}
\value{
A \code{multiOmicDataSet} object.
}
\description{
Read a multiOmicDataSet from disk
}
