% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOmicDataSet.R
\name{write_multiOmicDataSet}
\alias{write_multiOmicDataSet}
\title{Write a multiOmicDataSet to disk as RDS}
\usage{
write_multiOmicDataSet(moo, filepath = "moo.rds")
}
\arguments{
\item{moo}{multiOmicDataSet object to serialize.}

\item{filepath}{output path for RDS file.}
}
\value{
Invisibly returns \code{filepath}.
}
\description{
Write a multiOmicDataSet to disk as RDS
}
